pyproject.toml
[project]
name = "sparse-attention-zoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.1,<5",
    "transformers>=4.57.0,<5",
    'triton @ git+https://github.com/openai/triton.git@v3.4.0#egg=triton; sys_platform == "linux" and platform_machine == "aarch64"',
    'triton==3.4.0; sys_platform == "linux" and platform_machine == "x86_64"',
    "wandb>=0.22.2",
    "numpy<2",
    "sentencepiece>=0.2.1",
    "python-dotenv>=1.1.1",
]

# [tool.uv.sources]
# torch = [
#   # CUDA 12.9 for both x86_64 AND aarch64
#   {index = "pytorch-cu129", marker = 'sys_platform == "linux"'},
# ]

[tool.uv]
environments = [
  "sys_platform == 'linux' and platform_machine == 'x86_64'",
  "sys_platform == 'linux' and platform_machine == 'aarch64'"
]
override-dependencies = [
    "torch; sys_platform == 'never'", # so as to use the one installed in the container version 25.06
]
dependency-metadata = [
  {name = "torch", version = "2.8.0", requires-dist = ["numpy<2"]},
]
